fix: extensible cli interface
